Refactor project viewmodel 7600886583276935262

+++ app/src/main/java/com/facelapse/app/data/local/entity/ProjectEntity.kt
@@ -10,6 +10,6 @@ data class ProjectEntity(
     val name: String,
     val createdAt: Long = System.currentTimeMillis(),
     val isDateOverlayEnabled: Boolean = true,
-    val fps: Int = 10,
+    val fps: Float = 10f,
     val exportAsGif: Boolean = false
 )
